requirements.txt
mecab-python3==1.0.3
discord==1.0.1
discord.py==1.7.2
pydub==0.25.1
python-dotenv==0.17.1